split trainer tests
